refactor: dashboard structure and preview panel separation
